replace proportion plot by "who owns Britain"
