--- test/data/police-deaths.csv
@@ -1,5 +1,0 @@
-race,police,population
-Black,44,13
-Hispanic,32,12
-White,18,60
-All other races,6,15
