+++ test/data/wealth-britain.csv
@@ -1,0 +1,5 @@
+age,population,wealth
+16-34’s,30,3
+35-54’s,33,32
+55-74’s,28,52
+Over 75’s,9,13
